# swagger/MarketLocalShirtsAPI.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MarketLocalShirts API",
    "version": "1.0.0",
    "description": "API REST de MarketLocalShirts. Catálogo de camisas, autenticación JWT, usuarios y pedidos.\n\n1. POST /api/auth/login\n2. Copie el token\n3. Pulse Authorize y péguelo (sin la palabra Bearer)\n4. Pruebe el resto de endpoints."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor local de desarrollo"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT de POST /api/auth/login. Pegue solo el token."
      }
    },
    "parameters": {
      "page": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "default": 0 }
      },
      "size": {
        "name": "size",
        "in": "query",
        "schema": { "type": "integer", "default": 20 }
      },
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      }
    },
    "schemas": {
...
      "UsuarioGuardar": {
        "type": "object",
        "properties": {
          "nombre": { "type": "string", "example": "Valeria Admin" },
          "correo": { "type": "string", "example": "admin@marketlocalshirts.com" },
          "telefono": { "type": "string", "example": "7777-7777" },
          "contrasena": { "type": "string", "example": "Admin123" },
          "rol": { "type": "string", "example": "CLIENTE" }
        }
      },
      "UsuarioModificar": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "nombre": { "type": "string" },
          "correo": { "type": "string" },
          "telefono": { "type": "string" },
          "contrasena": { "type": "string" },
          "rol": { "type": "string", "example": "ADMIN" },
          "activo": { "type": "boolean" }
        }
      },
      "UsuarioSalida": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "nombre": { "type": "string" },
          "correo": { "type": "string" },
          "telefono": { "type": "string" },
          "rol": { "type": "string" },
          "activo": { "type": "boolean" }
        }
      },
...
      "CategoriaSalida": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "nombre": { "type": "string" },
          "descripcion": { "type": "string" }
        }
      },
...
      "MarcaSalida": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "nombre": { "type": "string" },
          "descripcion": { "type": "string" }
        }
      },
...
      "CamisaSalida": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "nombre": { "type": "string" },
          "descripcion": { "type": "string" },
          "precio": { "type": "number" },
          "stock": { "type": "integer" },
          "imagenUrl": { "type": "string" },
          "talla": { "type": "string" },
          "color": { "type": "string" },
          "categoria": { "$ref": "#/components/schemas/CategoriaSalida" },
          "marca": { "$ref": "#/components/schemas/MarcaSalida" }
        }
      },
      "DetallePedidoGuardar": {
        "type": "object",
        "properties": {
          "camisaId": { "type": "integer", "example": 1 },
          "cantidad": { "type": "integer", "example": 2 }
        }
      },
      "PedidoGuardar": {
        "type": "object",
        "properties": {
          "usuarioId": { "type": "integer", "example": 1 },
          "detalles": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/DetallePedidoGuardar" }
          }
        }
      },
      "PedidoModificar": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "estado": { "type": "string", "example": "CONFIRMADO" }
        }
      },
      "DetallePedidoSalida": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "camisa": { "$ref": "#/components/schemas/CamisaSalida" },
          "cantidad": { "type": "integer" },
          "precioUnitario": { "type": "number" },
          "subtotal": { "type": "number" }
        }
      },
      "PedidoSalida": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "usuario": { "$ref": "#/components/schemas/UsuarioSalida" },
          "fecha": { "type": "string", "format": "date-time" },
          "total": { "type": "number" },
          "estado": { "type": "string" },
          "detalles": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/DetallePedidoSalida" }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/usuarios": {
      "get": {
        "tags": ["Usuarios"],
        "summary": "Listar usuarios",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/UsuarioSalida" }
                }
              }
            }
          },
          "204": { "description": "Sin datos" }
        }
      },
      "post": {
        "tags": ["Usuarios"],
        "summary": "Crear usuario (admin)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UsuarioGuardar" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UsuarioSalida" }
              }
            }
          }
        }
      }
    },
    "/api/usuarios/paginado": {
      "get": {
        "tags": ["Usuarios"],
        "summary": "Listar usuarios paginados",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/size" }
        ],
        "responses": {
          "200": { "description": "Página" },
          "204": { "description": "Sin datos" }
        }
      }
    },
    "/api/usuarios/correo/{correo}": {
      "get": {
        "tags": ["Usuarios"],
        "summary": "Buscar usuario por correo",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "correo",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UsuarioSalida" }
              }
            }
          },
          "404": { "description": "No existe" }
        }
      }
    },
    "/api/usuarios/{id}": {
      "get": {
        "tags": ["Usuarios"],
        "summary": "Obtener usuario por ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "responses": {
          "200": {
            "description": "Encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UsuarioSalida" }
              }
            }
          },
          "404": { "description": "No existe" }
        }
      },
      "put": {
        "tags": ["Usuarios"],
        "summary": "Editar usuario, rol o estado",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UsuarioModificar" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Actualizado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UsuarioSalida" }
              }
            }
          },
          "404": { "description": "No existe" }
        }
      },
      "delete": {
        "tags": ["Usuarios"],
        "summary": "Eliminar usuario",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "responses": {
          "200": { "description": "Eliminado" }
        }
      }
    },
    "/api/pedidos": {
      "get": {
        "tags": ["Pedidos"],
        "summary": "Listar todos los pedidos",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/PedidoSalida" }
                }
              }
            }
          },
          "204": { "description": "Sin datos" }
        }
      },
      "post": {
        "tags": ["Pedidos"],
        "summary": "Confirmar pedido (checkout)",
        "description": "Valida stock y descuenta inventario. Si no hay existencias responde 400.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PedidoGuardar" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido confirmado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PedidoSalida" }
              }
            }
          },
          "400": { "description": "Stock insuficiente o datos inválidos" }
        }
      }
    },
    "/api/pedidos/paginado": {
      "get": {
        "tags": ["Pedidos"],
        "summary": "Listar pedidos paginados",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/size" }
        ],
        "responses": {
          "200": { "description": "Página" },
          "204": { "description": "Sin datos" }
        }
      }
    },
    "/api/pedidos/usuario/{usuarioId}": {
      "get": {
        "tags": ["Pedidos"],
        "summary": "Historial de pedidos de un usuario",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "usuarioId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/PedidoSalida" }
                }
              }
            }
          },
          "204": { "description": "Sin datos" },
          "403": { "description": "No puede ver pedidos de otro usuario" }
        }
      }
    },
    "/api/pedidos/{id}": {
      "get": {
        "tags": ["Pedidos"],
        "summary": "Obtener pedido por ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "responses": {
          "200": {
            "description": "Encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PedidoSalida" }
              }
            }
          },
          "404": { "description": "No existe" }
        }
      },
      "put": {
        "tags": ["Pedidos"],
        "summary": "Actualizar estado del pedido",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PedidoModificar" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Actualizado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PedidoSalida" }
              }
            }
          },
          "404": { "description": "No existe" }
        }
      },
      "delete": {
        "tags": ["Pedidos"],
        "summary": "Eliminar pedido",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "responses": {
          "200": { "description": "Eliminado" }
        }
      }
    }
  }
}